Add script to supervise initctl
This adds a script to wrap around the calls made to "initctl start $app",
"initctl restart $app" and "initctl reload $app" as part of an
application deployment. This script adds the ability to monitor that
these calls result in a new running process with any old processes ending.
It is only intended for apps that expose their process id in a pid file
at /var/run/$app/app.pid.

To start an app `sudo govuk_supervised_initctl start $app` is run. This
will run `initctl start` for the particular app and check that after a
short wait (10 seconds) that the pidfile for the app is a running
process. Prior to starting the app the wrapper will confirm if the
process is already running. If it turns out that upstart thinks a
process is running when the pid isn't actually running, the wrapper will
execute `initctl stop` to avoid getting stuck in a loop where this
wrapper and upstart disagree on a processes state (this can happen for
unicorn herder [1] managed GOV.UK apps that encounter syntax errors).
The inverse can also occur where a process is running but upstart is not
aware of it, however this is much less likely to occur, as it can't be
caused by the deployed application, so a simpler exit-with-error strategy
is used.

To restart an app `sudo govuk_supervised_initctl restart $app` is run.
This will check that a new process is started and that the old process
has stopped. This wraps around the `initctl restart` command, with a
similar approach to `start` where there is a 10 second wait before the
processes are checked. In order to run this command the old process
will need to be already running, therefore this call should be combined with
`govuk_supervised_initctl start` (for example: `sudo govuk_supervised_initctl
start $app || sudo govuk_supervised_initctl restart $app`)

To reload an app this script ports the behaviour from
`govuk_unicorn_reload` and supersedes that script (once this change is
applied I'll remove that from this repo). Compared to the start and
restart scripts this is a more involved process of running a loop for 3
minutes while waiting for the evidence that a new process is running and
the old process has stopped. There is more information on that script in
it's introductory commit message [2].

---

The motivation to add this script is to support continuous deployment on
GOV.UK. In particular, the need to know that a service is successfully
running after a restart [3]. I felt the most consistent way this could
be solved was by applying a wrapper command rather than modifying the
deployment scripts for services specifically. This wrapper is also
helpful for the other GOV.UK app services as it will fail a
deployment if these fail to start which will mean we know of a problem
sooner. This wrapper feels like something that shouldn't really be
needed for processes nicely integrated with upstart and, ideally, you'd
want upstart to take care of these concerns. However, this acts as a
mechanism to paper over the cracks in GOV.UK's integration (or my lack
of upstart knowledge, you decide).

I expect this wrapper to be applied by changing the Capistrano
deployment scripts [4] from:

```
if fetch(:perform_hard_restart, false)
  run "sudo initctl start #{application} 2>/dev/null || sudo initctl restart #{application}"
else
  run "sudo initctl start #{application} 2>/dev/null || sudo govuk_unicorn_reload #{application}"
end
```
to:

```
if fetch(:perform_hard_restart, false)
  run "sudo govuk_supervised_initctl start #{application} || sudo govuk_supervised_initctl restart #{application}"
else
  run "sudo govuk_supervised_initctl start #{application} || sudo govuk_supervised_initctl reload #{application}"
end
```

As this will need to be run via sudo I have added this script to the
list of executables the deploy user can run as sudo.

[1]: https://github.com/gds-operations/unicornherder
[2]: https://github.com/alphagov/govuk-puppet/commit/a3b33f09fc0fda565581dc593df1f98c08a9167b
[3]: https://github.com/alphagov/govuk-rfcs/blame/master/rfc-128-continuous-deployment.md#L143
[4]: https://github.com/alphagov/govuk-app-deployment/blob/e3f0bb339b60021fb0663aa9868cede064e8c1cd/recipes/ruby.rb#L20-L27