Do not alert on app restarts per day
We should not alert on app restarts due to OOM.
We should alert on symptoms affecting end users,
since we'll need to alert on those anyway. This
is an interesting metric, but not necessarily a
cause for any issues, and not actionable unless
it is impacting users.

An alternative would be to modify the alert so
we hear only when we're restarting a lot (e.g.
changing the restart window we check to ~3min).
This is sub-optimal since we are assuming this
is the rate at which users would notice, and
therefore we should just alert on the symptom
users experience (e.g. 5xx rate).