Add production check for Whitehall unenqueued scheduled documents
In https://github.com/alphagov/whitehall/pull/6001, we added an endpoint
to check for the number of scheduled editions that have not been queued
in Sidekiq.

We already have this data included in the Whitehall app check, but this
is shown in Icinga for every host (rather than once per app) and
suggests the entire app is critical, rather than one part. This will be
removed later to ensure continuity.