Raise severity of the alert when S3 mirror hasn't updated in >24h.
If the mirror job has been failing for whole day, that's not something
we would want to ignore.

Also expand the description a bit to give people a somewhat better
chance of guessing what the alert is about.

Today, I discovered (while investigating a completely unrelated issue) that the
mirror job had been failing for three days (because someone had disabled
it for "testing" and then forgotten about it). This warning-severity
alert hadn't alerted 2nd-line to the problem because there are always
lots of warnings and they usually don't mean that anything important is
wrong. There are other jobs which have noisy "Freshness threshold
exceeded" alerts so folks are particularly accustomed to ignoring these.