Stop setting the router_nodes_class variable
This commit stops the setting of the `router_nodes_class` variable for the `draft_cache` and `router_backend` machine classes. Router is being changed to self-update routes from MongoDB rather than having Router API inform it when to update routes; no longer setting these variables in Puppet is part of a wider set of changes to allow this to happen.