Do not prompt user to confirm SSH host key
If the user hasn't connected to the instance before, they'll be
prompted to confirm the host key.  Even worse, if they *have*
connected to the instance befoe, but it's been recycled (and so has a
new host key), SSH will reject the connection!

Since these instances are (in principle) ephemeral, and not the
permanent VMs of previous Carrenza days, then the host key is
unimportant.

By skipping host key verification, this script is a bit less
interactive.