Fallback to master branch for app-deployment
The APP_DEPLOYMENT_BRANCH environment variable is not consistently set
for all Jenkins runs of this job (I think it ends up always set when a
deploy is done through the UI but not set when a job is triggered via
API).

This change falls back to the default master branch if the env var is
not set.