Upgrade to graphite-1.0.0
The accounts team are considering using SLOs.  Unfortunately, we're
using graphite-0.9.13, which is not only ancient, but it's also
missing a ton of useful functions.

For example, to get a series which is 0 when 90th percentile latency
is below 500ms and 1 otherwise is currently:

    transformNull(offset(scale(removeBelowValue(maxSeries(stats.timers.*.nginx_logs.account-api.time_request.upper_90),0.5),0),1))

Whereas with graphite-1.0.0 it'll be:

    isNonNull(removeBelowValue(maxSeries(stats.timers.*.nginx_logs.account-api.time_request.upper_90),0.5))

And if we want a cumulative sum of that series over the last 48
days... forget it.  The `movingSum` function was only added in
graphite-1.0.0.  The best we can do currently is graph a 48 day range
with `integral(...)` but only look at the final value.  Not great if
we want to see how the moving sum has changed.

graphite-1.0.0 is still old, but it's a fairly small jump which
doesn't introduce any backwards incompatibilities we'll have to
resolve (other than a database migration, copied from the 1.0.0
release notes), and it has all the functions we want.

Release notes:

0.9.14 (2015-07-11): https://graphite.readthedocs.io/en/latest/releases/0_9_14.html
0.9.15 (2015-11-27): https://graphite.readthedocs.io/en/latest/releases/0_9_15.html
0.9.16 (2017-04-12): https://graphite.readthedocs.io/en/latest/releases/0_9_16.html
1.0.0  (2017-04-12): https://graphite.readthedocs.io/en/latest/releases/1_0_0.html