Remove newline at the end of config.xml.erb
It appears to be that when Jenkins reloads its own configuration, it
writes the file out again but without the trailing new line. This means
when Puppet next runs, it detects that the file is different, writes the
file, and reloads the Jenkins config. The newline was added in
066b0694165954491cef2e36d8ff9d12b4c9eb7b and we have removed it before
in 29acd89ae71e964a3c518997b4325767e1a40a60 and
dae2ac8bf349ba4ed7beef02f11a59ba8d9644e0.

I'm not sure how we can set this up so we don't put the new line back by
accident again...

Because of this, Jenkins is being reloaded approximately every 30
minutes, and I think this is causing it to lose track of any jobs
which are dependant on another one finished. Most recently, we've
been seeing this with the `Deploy_App_Downstream` job causing the
continuous deployment pipeline to get stuck. Checking the Jenkins logs,
I found a job which had been stuck, and coincidentally, the Jenkins
config had been reloaded just a few minutes after the stuck job started.

See
https://trello.com/c/hiZUg0gC/1142-smokey-failures-long-timeouts-blocking-continuous-deployment
for more details on this failure.