Add push job to Staging backup for Imminence
The push job in Staging was missing for imminence meaning the [pull
job][1] in Integration did not have any data to pull. 3:25am (push job
start time) is clear of any other backup jobs starting in Staging on the
db_admin machine and is 35 minutes before the pull job in Integration is
triggered which should give enough time to backup to the Staging S3
bucket.

[1]: https://github.com/alphagov/govuk-puppet/blob/master/hieradata_aws/class/integration/mongo.yaml#L24-L34

Trello:
https://trello.com/c/SLRaPvyI/2460-data-sync-not-working-for-imminence-5