Enable data sync for Release
This allows us to test the app in Integration and Staging with the
latest data, including running Rake tasks.

I've set it up such that the sync runs at the following times:

- 0:15 -> Backup Production
- 0:30 -> Backup Staging
- 1:15 <- Restore Staging from Production
- 0:45 -> Backup Integration
- 1:30 <- Restore Integration from Staging

This allows 45 minutes to backup each database, and 1 hour for the
backup to be available in S3 to restore from, which I think is plenty of
time.