Add Terraform 0.11.15 to installed versions.
For some reason we have preinstall specific Terraform versions via
Puppet even though we're using tfenv to install them and we're
specifying versions in .terraform-version files in the govuk-aws
Terraform repo.

Just changing the version number in .terraform-version causes tfenv to
fail with a permissions error, so this change is definitely necessary
with the current setup.