Limit range of keepLastValue in Sidekiq latency alerts
This was introduced in b0b17be:

> Now that statsd isn't repeating the last know values, use the
> keepLastValue function in graphite to smooth out the series for the
> monitoring targets.

However, since Sidekiq stops reporting latency when the queue empties,
I don't think this is quite right.  We should only keep the last value
a few minutes (I've gone for 36 datapoints, or 3 minutes here) to
smooth out the graph, but we don't want to keep it for the full 24
hours.

Currently, we have an issue where the search-api bulk latency alert is
triggered every morning and then never goes away.  Why?  Because once
a day we trigger a large batch of bulk jobs, which then run to
completion.  Since there are no more bulk jobs running, Sidekiq stops
reporting the latency, and so the `keepLastValue` makes it look like
we're constantly running at the latency of the last-to-complete job -
even though that job may have completed hours ago and the queue has
been empty since!

We don't see this problem with the other latency alerts, because those
queues don't have the same usage pattern of hitting the alert
threshold and then not running any more jobs.