Clear CDN cache in all environments for emergency banner
When the emergency banner is added or removed, we currently only clear
the CDN cache in production. This is because the CDN was previously only
used for production.

When the emergency banner was deployed for the first time on 9 April
2021, the cache clearing failed due to an error in the script. This was
not picked up during the emergency banner drills on staging as the
script had never run.

Now all environments are behind a CDN, we can safely clear the cache in
all environments when deploying or removing the banner.