Set up RabbitMQ for account-api
We want account-api to listen to publishing updates from RabbitMQ so
that we can update a user's saved pages when content gets updated.

For example, if we want to show the titles of the pages a user has
saved, we could do that by:

1. Querying content-store every time the user requests the list
2. Doing (1), but caching the results for a while
3. Querying content-store when the page is first saved, and updating
   local state when updates occur.

We're going for (3).  (1) introduces O(n) extra network requests; (2)
does the same, unless we have a long cache time, in which case we risk
showing stale data.  (3) introduces one extra network request when we
save a page, but no additional requests when the list is retrieved
and, while at risk of stale state, we can bound that risk by adopting
an SLO if we observe problems.

This commit is based on the cache-clearing-service RabbitMQ config.