fix the update of python2 to 2.7.14
This is the 2nd attempt (1st attempt in #11122) to install python 2.7.14
on all machines since the default Trusty version of pip2 is not SNI
compliant and no longer works.

Futher improvement on #11122 are:
1. installing govuk_python earlier in the process before govuk_scripts

2. merge `modules/govuk_jenkins/manifests/packages/govuk_python.pp` into
`modules/govuk_python/manifests/init.pp` to remove redundancy.

3. link `/usr/bin/pip` to `/opt/python2.7/bin/pip`

4. python3.6 is installed on mapit instances using `govuk-python3` repo
rather than mistakenly the `govuk-python` repo. This is required since
installation of Python 2.7.14 is via `govuk-python` repo and we can't
have 2 co-existing versions of the same apt package.

Co-Authored-By: Chris Ashton <ChrisBAshton@users.noreply.github.com>