Change Environment Agency harvest job cron time
We've had a request to change the harvest time from 5pm to 3pm:
https://govuk.zendesk.com/agent/tickets/4603384

Given that these cron jobs are only in place because the automated
runs fail, there is no particular time of day we need to adhere to,
so it is fine to bring this forward a couple of hours. See the
original commit eead57eab4d784122e3087b8e0fe35aa16053a16.

NB our machines are on UTC time, so 14:00 UTC is 15:00 BST.