Use OAuth access token instead of service account
The GOOGLE_CREDENTIALS key we were using before was a long-lived
credential with access to administer DNS.

We try to avoid giving Jenkins long lived tokens with high levels of
privilege. There are a few reasons why it's bad:

- The token is set to expire in the year 9999. A disgruntled former
  employee could steal this token and use it later.
- Using the service account complicates auditing who made a change to
  DNS - it just shows up as Jenkins in the GCP audit logs.
- Jenkins itself could be compromised (e.g. due to a bad plugin). If it
  always has access to privileged tokens, that could lead to an
  escalation.

For AWS, we have a pattern of pasting short lived STS credentials into
Jenkins.

Now that we've updated the GCP terraform provider to a version which
supports GOOGLE_OAUTH_ACCESS_TOKEN we can do the same thing for Google.

This requires that everyone who needs to make changes to DNS has the
`gcloud` CLI installed.

If this is approved, further work will be needed to:

- [ ] Update the documentation on how to deploy DNS
- [ ] Update the on-call checklist to make it clear that you need to have `gcloud` set up
- [ ] Email all the engineers on 2ndline to make them aware of the requirement to have `gcloud` set up
- [ ] Update the GDS CLI (which has a function where it can fill in the form in Jenkins for you)

https://trello.com/c/LasnNuz4/561-rationalise-use-of-gcp-credentials-eg-for-dns-deployments