Add Icinga warning for Account SLO rates (##11171)
The GOV.UK Accounts team has implimented SLOs to monitor the reliability
(measured by "bad" HTTP status codes as a proportion of all HTTP status
codes), and speed (measured by request latency, 5s worth of requets
are under 850ms).

We aim to keep both within a 99% Service Level Objective (SLO). This
gives us an error budget we wish to maintain.

In this PR we establish some fairly crude warning alerts that will warn
to Icinga and the #govuk-account-tech Slack channel if it looks like a
we have exceeded a 1% error rate across the last rolling ten minutes.

Whilst not ideal, it gives us some early warning of something going
wrong, allowing us to better plan and check on recent depoys. The theory
is, that if the rate continued at an above 1% error rate we would likely
see our error budget erroded - so best to know about it early and act
sooner.

We've been following some of the recommended monitoring patterns from
the [Google SRE book][1].

In particular this alert corresponds to their Alert example 1: [Target Error Rate ≥ SLO Threshold][2].

Whilst we are more interested in some of the other burn rate alerts
(which we could set to the duration of a sprint for example), we have
hit constraints with our current Graphite monitoring setup.

We will continue to look for better monitoring, but other solutions may
have to involve either:

- significatly more complicated formula,
- wait until replatforming has finished.

You can find more detais about issued encountered in this doc [SLO
graphite monitoring snag][3] (ask GOV.UK Accounts team for access if you need it)

[1]: https://sre.google/workbook/
[2]: https://sre.google/workbook/alerting-on-slos/#1-target-error-rate-slo-threshold
[3]: https://docs.google.com/document/d/1ZH58XIcuTcU8XhohAbbd5zxwIqqKjNVdpYgE9wHyVRk