Fix Account SLOs need a critical threshold
In commit 4187ef608 looks like I made puppet really sad.

```
Error: Could not retrieve catalog from remote server: Error 400 on
SERVER: Must pass critical to
Icinga::Check::Graphite[check_slo_error_rate_10_min_accounts] at
/usr/share/puppet/production/current/modules/monitoring/manifests/checks/accounts_slos.pp:27

Warning: Not using cache on failed catalog Error: Could not retrieve
catalog; skipping run
```

I believe this is because i had the bright idea to only set this up as a
warning. Mostly because it's an alert as an SLO early warning, and
critical seemed a bit alarmist.

So I've added a critical status as if we exceed 2000% of an error budget
in a rolling 10 minute window. That'd be a pretty bad day, and is
generally not somthing we should be seeing much.

That should mean this only ever really appears as a warning on any
Icinga dashboard, and so should not bug 2ndline too much with criticals
for our SLOs.