Change account SLO metric to use a movingMedian over 30s
Using the total error rate,

    divideSeries(integral($errors),integral($all))

Is far too noisy, at least with the volume of traffic we get.  It
looks like every single slow request is triggering the alert, spamming
our channel.  We're nowhere near running out of error budget, so this
sensitivity is too much.

So to make it less noisy, I've changed the metric to use the average
SLO violation rate over 30 seconds.  This smooths out spikes, but will
still quickly detect a real issue.

I've also changed the critical threshold to 1, which improves the
graph scale.  So now we'll raise a critical if *every* request is
violating the SLO, whereas previously we wouldn't.  That seems a
pretty bad problem so I'm ok with that.