Drop all but the last 6 datapoints for the account SLO alerts
Our weird graphite check thing doesn't use the latest value of the
metric, it uses the mean.  So it's currently taking the mean of the
`movingMedian` values, which is not what we want.

So change it to still calculate over 10 minutes, so we get a nice
graph to look at, but for the purposes of alerting only look at the
last 30s.