Set up sidekiq for account-api
We're setting up sidekiq for account-api, which means we need to set
up redis env vars for it first.  When this commit has been deployed,
we can deploy account-api with the new worker code, and then deploy a
second puppet change to enable the worker.

This commit also sets up sidekiq-monitoring and the application
dashboard.  Those won't have any metrics to show until the worker is
enabled though.