fix deploy uid/gid for deploy user in mirrorer machine
Do not set recurse on the mirror directory as it consists of a hundreds of GBs
of small files and will cause puppet to be take a long time to finish
running and may cause the Mirrorer instance to fail first instantiation.
Instead we set the uid and gid of `deploy` to be the same always via hiera.