Add a global hieradata variable for deploy jenkins domain
This sets up a global hiera variable in the same vein as app_domain for
Jenkins deploy hostname.

This has been done to put a plaster over the oddity where we have
deploy.integration.publishing.service.gov.uk in one environment and
deploy.blue.<env>.govuk.digital in others.

This variable has been put in to avoid having to put conditionals in
every place where there is a link to this Jenkins.