Consolidate deploy Jenkins domain handling
This replaces the various approaches to set the domain of the deploy
instance of Jenkins with a single one: use the global variable of
deploy_jenkins_domain.

Doing this will correct the incorrect links that are in Icinga.
Currently there are 2 broken production links and 5 in integration.

This applies a consistent pattern so removes some of the conditionals
used to determine the hostname (as seen in Smokey or Athena Fastly Logs
Check).

I applied this uniformly without determining whether all of the
variables and jobs were needed. I could see some of the jobs (Check
GitHub IP ranges) were not running in any environment but still had an
open PR [1]. I wasn't too convinced Slack needs a build server variable
configured but didn't know enough to test it.

[1]: https://github.com/alphagov/govuk-puppet/pull/10381