Set CheckHostIP to no in reboot-cache-instances script
When running this script checking the known_hosts file can
cause the SSH connection to be lost if the IP address has changed
since last connecting.

Users of this script would manually remove the problematic line
from the known_hosts file, and then have to either re-run the script
or more likely have to manually investigate the current state and
continue making changes in the AWS UI.

This commit removes this check allowing the script to complete.

https://man.openbsd.org/ssh_config#CheckHostIP