Change `ensure` to "present" for CKAN sync, not "enabled"
We're getting an error about the passive check associated with this
job.  The error is confusing, it says that we can't specify more than
one of `content`, `source`, or `target` - but it's clear that it's
something to do with this job.

Since all the other jobs use `ensure: "present"` and not `ensure:
"enabled"`, I think that line is at fault here.