Run mongodb check with python2.7
We're getting an error in Icinga:

    No module named pymongo

This is because the deploy user runs this with python 3, but we only
have the module installed for python 2.7.

This has to be specifically the version of python 2.7 in /opt, `env
python2` doesn't find the module either - we have other checks which
use this path too (like a few of the check_aws_* ones), so it's not
particularly weird.