Fix failures in Icinga name uniqueness check
The name for some icinga check resources must be unique, There is
currently one monitoring machine, so the name of each check is unique.
However when provisioning a new instance of monitoring, the names are
no longer unique so puppet fails.

Some checks already have the hostname appended to make them unique,
however the test for this was broken. As the test has now been fixed, we
need to add hostnames in places where they were missed when the tests
were added.

Adding the hostname to all the others to ensure every check has a unique
name for each provisioned instance.