Run app healthchecks as part of CD pipeline
One of the consequences of the CD RFC [1].

Currently these are written as bespoke scenarios in Smokey [2],
which is more work and confuses the purpose of the smoke tests
and leads to duplicate Icinga alerts. This makes the healthchecks
part of the CD pipeline so we can remove the bespoke scenarios.

Some apps don't support a web-based healthcheck, such as Cache
Clearing Service. In such cases we can use an empty array of URLs
to represent this. Other apps use a custom (Contacts Admin) or
even multiple domains, which we need to set explicitly.

Note that, unlike the existing scenario for Whitehall [3], this
tests both the Admin and Frontend instances due to the distinct
functionality. Conversely, it's not worth testing draft instances
of frontend apps as they behave the same as non-draft instances.

[1]: https://github.com/alphagov/govuk-rfcs/blob/main/rfc-128-continuous-deployment.md#healthy-deployments
[2]: https://github.com/alphagov/smokey/blob/cae82f827db5153f5545c342bd0f6ff19289e189/features/authenticating_proxy.feature
[3]: https://github.com/alphagov/smokey/blob/cae82f827db5153f5545c342bd0f6ff19289e189/features/whitehall.feature#L37-L41