Remove hostname from Icinga checks that run on monitoring
These checks only need to run on the current monitoring class machine,
so they do not need a unique check name. Otherwise another monitoring
instance will attempt to check these checks on other monitoring
machines, as well as it's own version of these checks.

This prevents Puppet running as the other machine's checks are not
accessible.