Add test for incorrectly added hostname in Icinga check title
Adds a check for the presence of the hostname in Incinga checks that run
on the `monitoring` machine class. These cannot be included as otherwise
Puppet will attempt to install multiple versions of the check on each
`monitoring` machine, which fails due to the other machine instance(s)
not being accessible to the newly provisoned machine.