Add Alex Whitehead-Smith
Add myself to the integration environment.

Note that I've created everything under my old surname (Smith). This is
so my SSH username matches the username on my GDS laptop. The email
address given here is an alias to the one on my Google account.

Hopefully this doesn't cause problems with the details in the [user monitor](https://github.com/alphagov/govuk-user-reviewer/pull/568).
If it does, I'll update that as well so it's all the same.