Give account-api a bearer token for publishing-api
So far we've used special-route-publisher to publish the
account-related routes (/sign-in, /sign-in/callback, /sign-out, and
/account/home).

These pages shouldn't appear in search, so the special route format
works for them.

But now we want to publish a "sorting hat" page---a routing page to
other government accounts---and this *should* appear in search.  We
can't use a special route for this.  We've decided to publish it as a
help page (as that's the closest matching format), but with some
custom behaviour which makes it unsuitable to publish through
mainstream publisher.

So we'll publish it directly from account-api, which means account-api
needs a bearer token for the publishing-api.