Update styling for GOV.UK Jenkins branding
Since we upgraded Jenkins the styling has been a bit janky. This applies
some rather crude styling fixes that sorts this out.

This code is totally coupled to Jenkins and will likely go a bit janky
again next time Jenkins changes styles. This should be ok though as this
happens so infrequently.

A bunch of the old styles here were no longer having an affect so they
have been removed.