Use govuk_personalisation in email-alert-frontend
In https://github.com/alphagov/email-alert-frontend/pull/1135 we switched
the email-alert-frontend to use `govuk_personalisation` when finding the
URL to redirect users to when changing their email address.

Set the required environment variables for this to work, and remove the
old ones. Note these are currently only set for integration. When we want
roll out changes further we'll need to update staging and production too.