Support play.http.secret and play.crypto.secret
We're [upgrading from Play 2.5 to Play 2.6](https://github.com/alphagov/licensify/pull/510),
and the config flag for the session cookie secret has changed its name.

In this commit we're setting both the old name and the new name to the
same value, so it will be forwards and backwards compatible.

Once the new version of Licensify is released we should remove the
play.crypto.secret value.