Deploy Static when adding/removing the emergency banner
Rather than clear caches when deploying or removing the emergency banner,
we will instead reduce the TTL of GOV.UK pages to 5 minutes, and wait for
the caches to expire - proposed in RFC 144 [1].

Sadly Static's cache does not expire automatically. Currently we
clear the Static caches with a script that removed the cached files
from the frontend machines.
This commit removes the clear-template-cache script and instead deploys
Static when adding/removing the banner, which will purge the template cache
and cause the templates to be regenerated.

[1]: https://github.com/alphagov/govuk-rfcs/pull/144